udevd: add -x/--exit-timeout option
On some systems with lots of devices, device probing for certain
drivers can take a very long time. If systemd-udevd detects a timeout
and kills the worker running modprobe, some devices will not be probed,
or end up in unusable state. The --event-timeout option can be used
to modify the maximum time spent in an uevent handler. But if systemd-udevd
exits, it uses a different timeout, so far hard-coded to 30s, and kills
all workers after this timeout expires. In practice, this may lead to
events being killed after significantly less time than specified with
the event-timeout, even if the default of 180s is used. This is
particularly significant during initrd processing, where systemd-udevd
will be stopped by systemd when initrd-switch-root.target is about
to be isolated, which usually happens quickly after finding and mounting
the root FS.

Fix this by making the exit timeout configurable.