udev: cleanup old symlink after creating new ones
If a high-prio device is processed which has remaining symlinks
from a previous database entry (which may be an entry that has survived
switching root because of the db_persist property), don't clean up
existing links before determining the current ones. Otherwise there
will be a time window in which the symlink points to a wrong,
lower-priority device, and accessing it may cause errors.

Therefore, create the new devlinks first, and remove only those that
belonged to the old device but not the new one.